requirements.txt
aioredis==1.3.1
amqp==5.0.9
appdirs==1.4.3
asgiref==3.4.1
asn1crypto==1.4.0
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
CacheControl==0.12.6
celery==5.2.3
certifi==2019.11.28
cffi==1.14.5
channels==3.0.4
channels-redis==3.3.1
chardet==3.0.4
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
constantly==15.1.0
contextlib2==0.6.0
cryptography==3.4.7
daphne==3.0.2
Deprecated==1.2.13
distlib==0.3.4
distro==1.4.0
Django==3.0.3
django-bootstrap4==2.2.0
django-extensions==3.1.5
djangorestframework==3.11.1
factory-boy==3.0.1
Faker==11.1.0
filelock==3.4.2
gunicorn==20.1.0
hiredis==2.0.0
html5lib==1.0.1
hyperlink==21.0.0
idna==2.8
incremental==21.3.0
ipaddr==2.2.0
kombu==5.2.3
lockfile==0.12.2
msgpack==1.0.3
packaging==21.3
pbr==5.8.0
pep517==0.8.2
Pillow==9.0.1
platformdirs==2.4.1
progress==1.5
prompt-toolkit==3.0.28
psycopg2==2.9.2
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==21.0.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytoml==0.1.21
pytz==2021.3
redis==4.1.4
requests==2.22.0
retrying==1.3.3
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
stevedore==3.5.0
text-unidecode==1.3
toml==0.10.2
Twisted==21.2.0
txaio==21.2.1
urllib3==1.25.8
vine==5.0.0
virtualenv==20.11.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
wrapt==1.13.3
zope.interface==5.4.0
